Module turbopack_node::evaluate

Structs

BuildDependencyIssue

An issue that occurred while evaluating node code.

BuildDependencyIssueVc

Vc for BuildDependencyIssue

EvaluateEmittedErrorIssue

EvaluateEmittedErrorIssueVc

Vc for EvaluateEmittedErrorIssue

EvaluationIssue

An issue that occurred while evaluating node code.

EvaluationIssueVc

Vc for EvaluationIssue

JavaScriptValueVc

Vc for JavaScriptValue

Enums

JavaScriptValue

Functions

evaluate

Pass the file you cared as runtime_entries to invalidate and reload the evaluated result automatically.

evaluate_inline

Pass the file you cared as runtime_entries to invalidate and reload the evaluated result automatically.

get_evaluate_pool

Pass the file you cared as runtime_entries to invalidate and reload the evaluated result automatically.

get_evaluate_pool_inline

Pass the file you cared as runtime_entries to invalidate and reload the evaluated result automatically.

Type Definitions

BuildDependencyIssueReadRef

see [turbo_tasks::ReadRef]

EvaluateEmittedErrorIssueReadRef

see [turbo_tasks::ReadRef]

EvaluationIssueReadRef

see [turbo_tasks::ReadRef]

JavaScriptValueReadRef

see [turbo_tasks::ReadRef]