Module turbopack_node::render::rendered_source

Structs

NodeRenderContentSource

see create_node_rendered_source

NodeRenderContentSourceVc

Vc for NodeRenderContentSource

Functions

create_node_rendered_source

Creates a content source that renders something in Node.js with the passed entry when it matches a path_regex. Once rendered it serves all assets referenced by the entry that are within the server_root. It needs a temporary directory (intermediate_output_path) to place file for Node.js execution during rendering. The chunking_context should emit to this directory.

create_node_rendered_source_inline

Creates a content source that renders something in Node.js with the passed entry when it matches a path_regex. Once rendered it serves all assets referenced by the entry that are within the server_root. It needs a temporary directory (intermediate_output_path) to place file for Node.js execution during rendering. The chunking_context should emit to this directory.

Type Definitions

NodeRenderContentSourceReadRef

see [turbo_tasks::ReadRef]